[ASV-1399] Moved AppLovin custom native event adapters to a mopub package
